+++ packages/design-system-docs/src/pages/components/FilterChip/_FilterChip.docs.scss
@@ -48,6 +48,7 @@ Style guide: components.filterchip.react
 ### Customization
 
 The following Sass variables can be overridden to the filter chip:
+- `$filter-chip-icon-container-size` - Height / Width for the close icon container
 - `$filter-chip-icon-size` - Height / Width for the close icon in the filter chip
 - `$filter-chip-text-color` - Text color for the filter chip
 - `$filter-chip-background-color` - Background color for the filter chip
